event.js 파일분리,  검색 시 이벤트발생함수 중복코드추출
event.js 파일 aside,list부분 이동 후
검색 시 이벤트발생함수 중복코드 추출, 메서드 호출
